quan_ly_dich_vu_xe/requirements.txt
mysql-connector-python==8.0.33
Pillow==10.0.0
reportlab==4.0.4